+++ chainclient/broadcast_tx.go
@@ -1,138 +1,116 @@
-package client
+package chainclient
 
 import (
 	"fmt"
+	"log"
+	"os"
 
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/tx"
-	"github.com/spf13/pflag"
+	cometrpc "github.com/cometbft/cometbft/rpc/client/http"
+	clitx "github.com/cosmos/cosmos-sdk/client/tx"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/vitwit/avail-da-module/types"
 
-	// tmjson "github.com/tendermint/tendermint/libs/json"
-	sdk "github.com/cosmos/cosmos-sdk/types"
+	"path/filepath"
+
+	"github.com/cosmos/cosmos-sdk/codec"
+	"github.com/cosmos/cosmos-sdk/crypto/keyring"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 )
 
-func (c *ChainClient) BroadcastTx(msg types.MsgSubmitBlobRequest, fromName string, fromAddr sdk.AccAddress) error {
-	fmt.Println("from name and from address.........", fromName, fromAddr)
-
-	clientCtx, err := c.BuildClientCtx(fromName, fromAddr)
+func GetBinPath() string {
+	homeDir, err := os.UserHomeDir()
 	if err != nil {
-		return err
+		log.Fatal(err)
 	}
 
-	flags := *pflag.NewFlagSet("my-flags", pflag.ContinueOnError)
-	fmt.Println("new flagssssss.......", flags)
+	availdHomePath := filepath.Join(homeDir, ".availsdk")
+	fmt.Println("availdHonmePath.......", availdHomePath)
+	return availdHomePath
+}
 
-	// err = clitx.GenerateOrBroadcastTxCLI(clientCtx, &flags, &msg)
-	txf, err := tx.NewFactoryCLI(clientCtx, &flags)
-	fmt.Println("here the eroor with txf....", txf, err)
+func ExecuteTX(ctx sdk.Context, msg types.MsgUpdateBlobStatusRequest, cdc codec.BinaryCodec) error {
+	// Define keyring and RPC client configuration
+
+	// homePath := "/home/vitwit/.availsdk"
+	homePath := GetBinPath()
+	key := os.Getenv("KEY")
+	fmt.Println("get key namee.........", key)
+	if key == "" { //TODO : remove this later
+		key = "alice"
+	}
+	keyName := key
+	rpcAddress := "http://localhost:26657"
+
+	// Create a keyring
+	kr, err := keyring.New(sdk.KeyringServiceName(), keyring.BackendTest, homePath, os.Stdin, cdc.(codec.Codec))
 	if err != nil {
-		return err
+		return fmt.Errorf("error creating keyring: %w", err)
 	}
 
-	err = tx.GenerateOrBroadcastTxWithFactory(clientCtx, txf, &msg)
+	// List all keys in the keyring
+	// keys, err := kr.List()
+	// if err != nil {
+	// 	fmt.Println("error listing keys:", err)
+	// }
+
+	info, err := kr.Key(keyName)
+	valAddr, err := info.GetAddress()
+	fmt.Println("after address................", valAddr)
+
+	// valAddr, err := sdk.AccAddressFromBech32(addr.String())
+	// fmt.Println("val addr, err..", valAddr, err, addr)
+
+	// fmt.Println("keysss........", keys)
+
+	// // Print out the keys
+	// for _, keyInfo := range keys {
+	// 	addr, err := keyInfo.GetAddress()
+	// 	fmt.Println("err..", err)
+	// 	fmt.Printf("Name: %s, Address: %s\n", keyInfo.Name, addr)
+	// }
+
+	// Create an RPC client
+	rpcClient, err := cometrpc.NewWithTimeout(rpcAddress, "/websocket", 3)
 	if err != nil {
-		fmt.Println("error insideeeeeeeeeeee............", err)
-		return err
+		return fmt.Errorf("error creating RPC client: %w", err)
+	}
+
+	// Create a new client context
+	clientCtx := NewClientCtx(kr, rpcClient, ctx.ChainID(), cdc, homePath, valAddr)
+
+	// Retrieve the validator address (replace with actual logic to get the address)
+	// valAddr, err = sdk.AccAddressFromBech32("cosmos1fhqer4tc50nut2evvnj6yegcah2yfu3s844n9a")
+	// if err != nil {
+	// 	return fmt.Errorf("error parsing validator address: %w", err)
+	// }
+
+	// Set the client context's from fields
+	clientCtx.FromName = keyName
+	clientCtx.FromAddress = valAddr
+
+	// Fetch account number and sequence from the blockchain
+	accountRetriever := authtypes.AccountRetriever{}
+	account, err := accountRetriever.GetAccount(clientCtx, valAddr)
+	if err != nil {
+		return fmt.Errorf("error retrieving account: %w", err)
+	}
+
+	fmt.Println("account details......", account.GetAccountNumber(), account.GetSequence())
+
+	// Set the correct account number and sequence
+	factory := NewFactory(clientCtx).
+		WithAccountNumber(account.GetAccountNumber()).
+		WithSequence(account.GetSequence())
+
+	// Create a transaction factory and set the validator address in the message
+	// factory := NewFactory(clientCtx)
+	msg.ValidatorAddress = valAddr.String()
+	// time.Sleep(10 * time.Second)
+
+	// Generate and broadcast the transaction
+	if err := clitx.GenerateOrBroadcastTxWithFactory(clientCtx, factory, &msg); err != nil {
+		return fmt.Errorf("error broadcasting transaction: %w", err)
 	}
 
 	return nil
 }
-
-// BuildClientCtx builds the context for the client
-func (c *ChainClient) BuildClientCtx(accountName string, accountAddress sdk.AccAddress) (client.Context, error) {
-	// info, err := c.clientCtx.Keyring.Key(accountName)
-	// if err != nil {
-	// 	return client.Context{}, err
-	// }
-	return c.clientCtx.
-		WithFromName(accountName).
-		WithFromAddress(accountAddress), nil
-}
-
-// // PrepareBroadcast performs checks and operations before broadcasting messages
-// func (c *ChainClient) PrepareBroadcast(msgs ...types.Msg) error {
-// 	// validate msgs
-// 	for _, msg := range msgs {
-// 		if err := msg.ValidateBasic(); err != nil {
-// 			return err
-// 		}
-// 	}
-
-// 	c.out.Reset()
-
-// 	return nil
-// }
-
-// // SignTx signs tx and return tx bytes
-// func (c *ChainClient) SignTx(fromName string, fromAddr types.AccAddress, clientCtx client.Context, msgs ...types.Msg) ([]byte, error) {
-// 	clientCtx, err := c.BuildClientCtx(fromName, fromAddr)
-// 	if err != nil {
-// 		return []byte{}, err
-// 	}
-
-// 	if err := c.PrepareBroadcast(msgs...); err != nil {
-// 		return []byte{}, err
-// 	}
-
-// 	flags := *pflag.NewFlagSet("my-flags", pflag.ContinueOnError)
-
-// 	txf, err := tx.NewFactoryCLI(clientCtx, &flags)
-// 	if err != nil {
-// 		return []byte{}, err
-// 	}
-
-// 	unsignedTx, err := tx.BuildUnsignedTx(txf, msgs...)
-// 	if err != nil {
-// 		return []byte{}, err
-// 	}
-
-// 	err = tx.Sign(txf, clientCtx.GetFromName(), unsignedTx, true)
-// 	if err != nil {
-// 		return []byte{}, err
-// 	}
-// 	return clientCtx.TxConfig.TxEncoder()(unsignedTx.GetTx())
-// }
-
-// // Broadcast directly broadcasts the messages
-// func (c *ChainClient) Broadcast(fromName string, fromAddr types.AccAddress, clientCtx client.Context, msgs ...types.Msg) (*types.TxResponse, error) {
-// 	clientCtx, err := c.BuildClientCtx(fromName, fromAddr)
-// 	if err != nil {
-// 		return &types.TxResponse{}, err
-// 	}
-
-// 	if err := c.PrepareBroadcast(msgs...); err != nil {
-// 		return &types.TxResponse{}, err
-// 	}
-
-// 	// broadcast tx.
-// 	if err := tx.BroadcastTx(clientCtx, c.factory, msgs...); err != nil {
-// 		return &types.TxResponse{}, err
-// 	}
-
-// 	// return c.handleBroadcastResult()
-// 	return &types.TxResponse{}, nil
-// }
-
-// // HandleBroadcastResult handles the result of broadcast messages result and checks if an error occurred
-// // func (c *ChainClient) handleBroadcastResult() (*types.TxResponse, error) {
-// // 	var out types.TxResponse
-// // 	if err := tmjson.Unmarshal(c.out.Bytes(), &out); err != nil {
-// // 		return &out, err
-// // 	}
-// // 	if out.Code > 0 {
-// // 		return &out, fmt.Errorf("tx error with code '%d' code: %s", out.Code, out.RawLog)
-// // 	}
-// // 	return &out, nil
-// // }
-
-// // BuildClientCtx builds the context for the client
-// func (c *ChainClient) BuildClientCtx(accountName string, accountAddress types.AccAddress) (client.Context, error) {
-// 	// info, err := c.clientCtx.Keyring.Key(accountName)
-// 	// if err != nil {
-// 	// 	return client.Context{}, err
-// 	// }
-// 	return c.clientCtx.
-// 		WithFromName(accountName).
-// 		WithFromAddress(accountAddress), nil
-// }
