+++ relayer/publish.go
@@ -2,16 +2,12 @@ package relayer
 
 import (
 	"fmt"
-	"os"
 
-	cometrpc "github.com/cometbft/cometbft/rpc/client/http"
-	clitx "github.com/cosmos/cosmos-sdk/client/tx"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/vitwit/avail-da-module/types"
 
 	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/cosmos-sdk/crypto/keyring"
-	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	dacli "github.com/vitwit/avail-da-module/chainclient"
 )
 
 // PostNextBlocks is called by the current proposing validator during PrepareProposal.
@@ -106,7 +102,7 @@ func (r *Relayer) postBlocks(ctx sdk.Context, blocks []int64, cdc codec.BinaryCo
 		)
 
 		// TODO : execute tx about failure submission
-		err = ExecuteTX(ctx, types.MsgUpdateBlobStatusRequest{
+		err = dacli.ExecuteTX(ctx, types.MsgUpdateBlobStatusRequest{
 			ValidatorAddress: sdk.AccAddress.String(proposer),
 			BlocksRange: &types.Range{
 				From: uint64(blocks[0]),
@@ -137,7 +133,7 @@ func (r *Relayer) postBlocks(ctx sdk.Context, blocks []int64, cdc codec.BinaryCo
 		fmt.Println("submit blocks msg.......", msg)
 
 		// TODO : execute tx about successfull submission
-		err = ExecuteTX(ctx, msg, cdc)
+		err = dacli.ExecuteTX(ctx, msg, cdc)
 		if err != nil {
 			fmt.Println("error while submitting tx...", err)
 		}
@@ -145,85 +141,100 @@ func (r *Relayer) postBlocks(ctx sdk.Context, blocks []int64, cdc codec.BinaryCo
 
 }
 
-func ExecuteTX(ctx sdk.Context, msg types.MsgUpdateBlobStatusRequest, cdc codec.BinaryCodec) error {
-	// Define keyring and RPC client configuration
+// var availdHomePath = xfilepath.JoinFromHome(xfilepath.Path("availsdk"))
 
-	homePath := "/home/vitwit/.availsdk"
-	keyName := "alice"
-	rpcAddress := "http://localhost:26657"
+// availdHomePath := filepath.Join(os.Getenv("HOME"), "availsdk")
 
-	// Create a keyring
-	kr, err := keyring.New(sdk.KeyringServiceName(), keyring.BackendTest, homePath, os.Stdin, cdc.(codec.Codec))
-	if err != nil {
-		return fmt.Errorf("error creating keyring: %w", err)
-	}
+// func GetBinPath() string {
+// 	homeDir, err := os.UserHomeDir()
+// 	if err != nil {
+// 		log.Fatal(err)
+// 	}
 
-	// List all keys in the keyring
-	// keys, err := kr.List()
-	// if err != nil {
-	// 	fmt.Println("error listing keys:", err)
-	// }
+// 	availdHomePath := filepath.Join(homeDir, ".availsdk")
+// 	fmt.Println("availdHonmePath.......", availdHomePath)
+// 	return availdHomePath
+// }
 
-	info, err := kr.Key(keyName)
-	// log.Println("uuu....", info, err)
-	fmt.Println("here error???", info == nil)
-	valAddr, err := info.GetAddress()
-	fmt.Println("after address................", valAddr)
+// func ExecuteTX(ctx sdk.Context, msg types.MsgUpdateBlobStatusRequest, cdc codec.BinaryCodec) error {
+// 	// Define keyring and RPC client configuration
 
-	// valAddr, err := sdk.AccAddressFromBech32(addr.String())
-	// fmt.Println("val addr, err..", valAddr, err, addr)
+// 	// homePath := "/home/vitwit/.availsdk"
+// 	homePath := GetBinPath()
+// 	fmt.Println("get key namee.........", os.Getenv("KEY"))
+// 	keyName := os.Getenv("KEY")
+// 	rpcAddress := "http://localhost:26657"
 
-	// fmt.Println("keysss........", keys)
+// 	// Create a keyring
+// 	kr, err := keyring.New(sdk.KeyringServiceName(), keyring.BackendTest, homePath, os.Stdin, cdc.(codec.Codec))
+// 	if err != nil {
+// 		return fmt.Errorf("error creating keyring: %w", err)
+// 	}
 
-	// // Print out the keys
-	// for _, keyInfo := range keys {
-	// 	addr, err := keyInfo.GetAddress()
-	// 	fmt.Println("err..", err)
-	// 	fmt.Printf("Name: %s, Address: %s\n", keyInfo.Name, addr)
-	// }
+// 	// List all keys in the keyring
+// 	// keys, err := kr.List()
+// 	// if err != nil {
+// 	// 	fmt.Println("error listing keys:", err)
+// 	// }
 
-	// Create an RPC client
-	rpcClient, err := cometrpc.NewWithTimeout(rpcAddress, "/websocket", 3)
-	if err != nil {
-		return fmt.Errorf("error creating RPC client: %w", err)
-	}
+// 	info, err := kr.Key(keyName)
+// 	valAddr, err := info.GetAddress()
+// 	fmt.Println("after address................", valAddr)
 
-	// Create a new client context
-	clientCtx := NewClientCtx(kr, rpcClient, ctx.ChainID(), cdc, homePath, valAddr)
+// 	// valAddr, err := sdk.AccAddressFromBech32(addr.String())
+// 	// fmt.Println("val addr, err..", valAddr, err, addr)
 
-	// Retrieve the validator address (replace with actual logic to get the address)
-	// valAddr, err = sdk.AccAddressFromBech32("cosmos1fhqer4tc50nut2evvnj6yegcah2yfu3s844n9a")
-	// if err != nil {
-	// 	return fmt.Errorf("error parsing validator address: %w", err)
-	// }
+// 	// fmt.Println("keysss........", keys)
 
-	// Set the client context's from fields
-	clientCtx.FromName = keyName
-	clientCtx.FromAddress = valAddr
+// 	// // Print out the keys
+// 	// for _, keyInfo := range keys {
+// 	// 	addr, err := keyInfo.GetAddress()
+// 	// 	fmt.Println("err..", err)
+// 	// 	fmt.Printf("Name: %s, Address: %s\n", keyInfo.Name, addr)
+// 	// }
 
-	// Fetch account number and sequence from the blockchain
-	accountRetriever := authtypes.AccountRetriever{}
-	account, err := accountRetriever.GetAccount(clientCtx, valAddr)
-	if err != nil {
-		return fmt.Errorf("error retrieving account: %w", err)
-	}
+// 	// Create an RPC client
+// 	rpcClient, err := cometrpc.NewWithTimeout(rpcAddress, "/websocket", 3)
+// 	if err != nil {
+// 		return fmt.Errorf("error creating RPC client: %w", err)
+// 	}
 
-	fmt.Println("account details......", account.GetAccountNumber(), account.GetSequence())
+// 	// Create a new client context
+// 	clientCtx := NewClientCtx(kr, rpcClient, ctx.ChainID(), cdc, homePath, valAddr)
 
-	// Set the correct account number and sequence
-	factory := NewFactory(clientCtx).
-		WithAccountNumber(account.GetAccountNumber()).
-		WithSequence(account.GetSequence())
+// 	// Retrieve the validator address (replace with actual logic to get the address)
+// 	// valAddr, err = sdk.AccAddressFromBech32("cosmos1fhqer4tc50nut2evvnj6yegcah2yfu3s844n9a")
+// 	// if err != nil {
+// 	// 	return fmt.Errorf("error parsing validator address: %w", err)
+// 	// }
 
-	// Create a transaction factory and set the validator address in the message
-	// factory := NewFactory(clientCtx)
-	msg.ValidatorAddress = valAddr.String()
-	// time.Sleep(10 * time.Second)
+// 	// Set the client context's from fields
+// 	clientCtx.FromName = keyName
+// 	clientCtx.FromAddress = valAddr
 
-	// Generate and broadcast the transaction
-	if err := clitx.GenerateOrBroadcastTxWithFactory(clientCtx, factory, &msg); err != nil {
-		return fmt.Errorf("error broadcasting transaction: %w", err)
-	}
+// 	// Fetch account number and sequence from the blockchain
+// 	accountRetriever := authtypes.AccountRetriever{}
+// 	account, err := accountRetriever.GetAccount(clientCtx, valAddr)
+// 	if err != nil {
+// 		return fmt.Errorf("error retrieving account: %w", err)
+// 	}
 
-	return nil
-}
+// 	fmt.Println("account details......", account.GetAccountNumber(), account.GetSequence())
+
+// 	// Set the correct account number and sequence
+// 	factory := NewFactory(clientCtx).
+// 		WithAccountNumber(account.GetAccountNumber()).
+// 		WithSequence(account.GetSequence())
+
+// 	// Create a transaction factory and set the validator address in the message
+// 	// factory := NewFactory(clientCtx)
+// 	msg.ValidatorAddress = valAddr.String()
+// 	// time.Sleep(10 * time.Second)
+
+// 	// Generate and broadcast the transaction
+// 	if err := clitx.GenerateOrBroadcastTxWithFactory(clientCtx, factory, &msg); err != nil {
+// 		return fmt.Errorf("error broadcasting transaction: %w", err)
+// 	}
+
+// 	return nil
+// }
